requirements.txt
spacy~=3.3.1
deepl~=1.11.0
requests~=2.28.1
pandas~=1.4.4
tqdm~=4.64.1
google-cloud-translate==2.0.1